Moving the removeMetrics() back to PerformanceAnalyzerMetrics class
