+++ src/main/java/org/opensearch/performanceanalyzer/reader_writer_shared/EventLogFileHandler.java
@@ -42,12 +42,8 @@ import java.nio.file.StandardOpenOption;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.message.ParameterizedMessage;
-import org.apache.logging.log4j.util.Supplier;
-import org.opensearch.performanceanalyzer.PerformanceAnalyzerApp;
 import org.opensearch.performanceanalyzer.core.Util;
 import org.opensearch.performanceanalyzer.metrics.PerformanceAnalyzerMetrics;
-import org.opensearch.performanceanalyzer.rca.framework.metrics.WriterMetrics;
 import org.opensearch.performanceanalyzer.reader.EventDispatcher;
 
 public class EventLogFileHandler {
@@ -172,7 +168,7 @@ public class EventLogFileHandler {
         }
     }
 
-    private void deleteFiles(long referenceTime, int purgeInterval) {
+    public void deleteFiles(long referenceTime, int purgeInterval) {
         LOG.debug("Starting to delete old writer files");
         File root = new File(metricsLocation);
         String[] children = root.list();
@@ -184,44 +180,10 @@ public class EventLogFileHandler {
             File fileToDelete = new File(root, child);
             if (fileToDelete.lastModified()
                     < PerformanceAnalyzerMetrics.getTimeInterval(referenceTime - purgeInterval)) {
-                removeMetrics(fileToDelete);
+                PerformanceAnalyzerMetrics.removeMetrics(fileToDelete);
                 filesDeletedCount += 1;
             }
         }
         LOG.debug("'{}' Old writer files cleaned up.", filesDeletedCount);
-    }
-
-    public static void removeMetrics(String keyPath) {
-        removeMetrics(new File(keyPath));
-    }
-
-    public static void removeMetrics(File keyPathFile) {
-        if (keyPathFile.isDirectory()) {
-            String[] children = keyPathFile.list();
-            if (children != null) {
-                for (String child : children) {
-                    removeMetrics(new File(keyPathFile, child));
-                }
-            }
-        }
-        try {
-            if (!keyPathFile.delete()) {
-                PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
-                        WriterMetrics.METRICS_REMOVE_FAILURE, "", 1);
-                LOG.debug("Purge Could not delete file {}", keyPathFile);
-            }
-        } catch (Exception ex) {
-            PerformanceAnalyzerApp.WRITER_METRICS_AGGREGATOR.updateStat(
-                    WriterMetrics.METRICS_REMOVE_ERROR, "", 1);
-            LOG.debug(
-                    (Supplier<?>)
-                            () ->
-                                    new ParameterizedMessage(
-                                            "Error in deleting file: {} for keyPath:{} with ExceptionCode: {}",
-                                            ex.toString(),
-                                            keyPathFile.getAbsolutePath(),
-                                            WriterMetrics.METRICS_REMOVE_ERROR.toString()),
-                    ex);
-        }
     }
 }
